requirements.txt
astroid==2.0.4
Django==1.9
django-bootstrap-form==3.4
django-datetime-widget==0.9.3
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
